// services/dto/cart-dto.ts
import type { Cart, CartItem, Ingredient, Product, ProductItem } from '@prisma/client'

export type CartItemDTO = CartItem & {
  productItem: ProductItem & { product: Product; ingredients: Ingredient[] }
  ingredients: Ingredient[]
}

export interface CartDTO extends Cart {
  items: CartItemDTO[]
}

export interface CreateCartItemValues {
  productItemId: number
  ingredients?: number[]
}